Fix TaskForm spec
