+++ frontend/src/app/tasks/hooks/useTask.ts
@@ -2,9 +2,10 @@ import { useMutation } from '@tanstack/react-query'
 import { TaskIntent } from '@/domain/Task'
 import { useAlert } from '@/ui/Alert/useAlert'
 import { createTask } from '@/infra/task/createTask'
-import { useClientFetch } from '@/infra/lib/apiClient'
+import { useClientFetch } from '@/infra/lib/useClientFetch'
 
 export const useAddTask = () => {
+  console.log('test')
   const apiClient = useClientFetch()
   const { showError, showSuccess } = useAlert()
